Create moving_target.py
Problem:
You are at the shooting gallery again, and you need a program that helps you keep track of moving targets. On the first line, you will receive a sequence of targets with their integer values, split by a single space. Then, you will start receiving commands for manipulating the targets until the "End" command. The commands are the following:
- "Shoot {index} {power}"
Shoot the target at the index if it exists by reducing its value by the given power (integer value). A target is considered shot when its value reaches 0.
Remove the target if it is shot.
- "Add {index} {value}"
Insert a target with the received value at the received index if it exists. If not, print: "Invalid placement!"
- "Strike {index} {radius}"
Remove the target at the given index (if such exist) and the ones before and after it depending on the radius. 
If any of the indices in the range is invalid, print: "Strike missed!" and skip this command.
- "End"
Print the sequence with targets in the following format:
"{target_1}|{target_2} … |{target_n}"

Input / Constraints:
- On the first line, you will receive the sequence of targets – integer values [1-10000].
- On the following lines, until the "End", you will be receiving the command described above – strings.
- There will never be a case when the "Strike" command would empty the whole sequence.